added factory and state

+++ src/Services/FullBoard.java
@@ -9,7 +9,7 @@ public class FullBoard extends ServicesDecorator {
     }
 
     @Override
-    public Float getCost() {
+    public double getCost() {
         return (float)400.0;
     }
 }
